+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_wrong_names.py
@@ -2,7 +2,6 @@
 
 import pytest
 
-from wemake_python_styleguide.constants import VARIABLE_NAMES_BLACKLIST
 from wemake_python_styleguide.violations.naming import (
     WrongVariableNameViolation,
 )